projects/requirements.txt
numpy~=2.2.2
pandas~=2.2.3
matplotlib~=3.10.0
seaborn~=0.13.2
statsmodels~=0.14.4
joblib~=1.4.2
scikit-learn~=1.6.1
pyyaml~=6.0
fastapi~=0.115.0
uvicorn~=0.32.0
pydantic~=2.10.0

# 05_agents — install Ollama separately: https://ollama.ai
mcp>=1.0.0
ollama>=0.3.0
